requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
elasticsearch==8.11.1
python-jose==3.3.0
httpx==0.26.0
certifi==2024.2.2 